Allow only self-hosted renovate to operate here
renovate app and MendRenovate are stubborn (surprise!) and they keep
trying to run if they get access to the repository and they find a
configuration file.

Add the necessary countermeasures to prevent that. Yes, countermeasures.
At this point renovate feels like something you have to fight against
and not something you work with.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>