feat: renovate-approve: add `force-merge` functionality
If a repository is configured to require review from CODEOWNERS, then
the approver's app approval will be useless. As apps cannot be added to
CODEOWNERS (https://github.com/orgs/community/discussions/23064), this
commit introduces a workaround: A toggle to tell the approver app to
also forcefully merge the PR.

For this to work, the application needs to be added to the bypass list
of the repository.