feat: presets/sm: add automerge rules for GHA, Go and Containers
This commit adds a first iteration of automerge rules for a restricted
subset of packages and ecosystems.

Specifically, this change will cause renovate to enable automerge via
GitHub for:
- Official and first-party GitHub Actions
- Toolchain, first-party, and `golang.org/x/` Go packages
- First-party container images

Where "first-party" means "Grafana".

In all cases, automerge will only be enabled for minor, patch, and
digest type updates, never for majors.

This rules will cause renovate to _request_ GitHub to automerge the PRs,
but it is not sufficient for the automerge to happen. For that, two
other requirements must be met:
- The PR must be approved by a reviewer with write access (which may be
the auto-approver bot, if enabled)
- The PR must pass all CI/CD checks